Moved all our TODOs into a separate notes file. I also consolidated all our actuator classes into a single "Devices" file.
