+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Notes and TODOs
@@ -1,0 +1,7 @@
+//TODO, implement all our external libraries and functionality.
+
+//TODO, add 2 generic autos for preloads and shooting and stuff
+
+//TODO, Implement controls graph
+
+//DONE, Consolidate the generic actuator functionalities folder into one file
